==> notebooks/01-Preprocess.ipynb <==
Extract the subset of the data we want to look at.

import numpy as np
import matplotlib as mpl
import matplotlib.pyplot as plt
import pandas as pd
import seaborn as sns

# Identify columns (2020 data)

Use the Alaska data to identify the colums we want.

def read_state_2020_df(state):
    df = pd.read_excel(f"../data/rwjf_2020/{state}.xlsx", "Additional Measure Data", header=1)
    # standardize the column name of Borough/County/Parish
    df = df.rename(columns={'Borough': 'Division', 'County': 'Division', 'Parish': 'Division'})
    return df

df = read_state_2020_df("Alaska")

df.head()

df.columns

le_cols = [c for c in df.columns if c.startswith('Life Expectancy (') and c.endswith(')')]

pop_cols = ['Population', 
            '# Black', '% Black', 
            '# American Indian & Alaska Native', '% American Indian & Alaska Native', 
            '# Asian', '% Asian',
            '# Native Hawaiian/Other Pacific Islander',
            '% Native Hawaiian/Other Pacific Islander', 
            '# Hispanic', '% Hispanic',
            '# Non-Hispanic White', '% Non-Hispanic White',
            '% Female']

income_cols = [c for c in df.columns if "Income" in c]

cols = ['FIPS', 'State', 'Division', 'Life Expectancy']
cols.extend(le_cols)
cols.extend(pop_cols)
cols.extend(income_cols)

Here are the columns

cols

# Subset data

Create a table with these columns for all states and territories.

states = [
'Alabama',
'Alaska',
'Arizona',
'Arkansas',
'California',
'Colorado',
'Connecticut',
'Delaware',
'District_of_Columbia',
'Florida',
'Georgia',
'Hawaii',
'Idaho',
'Illinois',
'Indiana',
'Iowa',
'Kansas',
'Kentucky',
'Louisiana',
'Maine',
'Maryland',
'Massachusetts',
'Michigan',
'Minnesota',
'Mississippi',
'Missouri',
'Montana',
'Nebraska',
'Nevada',
'New_Hampshire',
'New_Jersey',
'New_Mexico',
'New_York',
'North_Carolina',
'North_Dakota',
'Ohio',
'Oklahoma',
'Oregon',
'Pennsylvania',
'Rhode_Island',
'South_Carolina',
'South_Dakota',
'Tennessee',
'Texas',
'Utah',
'Vermont',
'Virginia',
'Washington',
'West_Virginia',
'Wisconsin',
'Wyoming'
]

subsets = []
for state in states:
    print(f"Processing {state}")
    df = read_state_df(state)
    subsets.append(df[cols])

len(subsets)

# Export data

First the division-level data

df_2020 = pd.concat(subsets)

# Look at only Borough/County/Parish data
df_2020 = df_2020.dropna(subset='Division')

df_2020.head()

df_2020.to_csv("../data/rwjf/division_2020.csv", index=False)

Then the state-level data

df_2020_states = pd.concat(subsets)
df_2020_states = df_2020_states[pd.isna(df_2020_states['Division'])]
df_2020_states.head()

df_2020_states.to_csv("../data/rwjf/state_2020.csv", index=False)

# Identify columns (2022 data)

Now do the same for the 2022 data

def read_state_2022_df(state):
    df = pd.read_excel(f"../data/rwjf_2022/{state}.xlsx", "Additional Measure Data", header=1)
    # standardize the column name of Borough/County/Parish
    df = df.rename(columns={'Borough': 'Division', 'County': 'Division', 'Parish': 'Division'})
    return df

df = read_state_2022_df("Alaska")

df.head()

Let us check that the columns from the 2020 data are all still here

found_cols = []
missing_cols = []
for c in cols:
    if c not in df.columns:
        missing_cols.append(c)
    else:
        found_cols.append(c)
missing_cols

There are a few columns we need to fix up

It looks like 'White' was renamed to 'white' and 'Female' was changed to 'female'. Let us reverse that so the format is consistent to work with.

rename_dict = {c.replace('White', 'white'): c for c in missing_cols}
rename_dict['% female'] = '% Female'
tdf = df.rename(columns=rename_dict)

found_cols = []
missing_cols = []
for c in cols:
    if c not in tdf.columns:
        missing_cols.append(c)
    else:
        found_cols.append(c)
missing_cols

Ok, let us work with this.

def read_state_2022_df(state):
    df = pd.read_excel(f"../data/rwjf_2022/{state}.xlsx", "Additional Measure Data", header=1)
    # standardize the column name of Borough/County/Parish
    df = df.rename(columns={'Borough': 'Division', 'County': 'Division', 'Parish': 'Division'})
    df = df.rename(columns=rename_dict)
    return df

# Subset data

Create a table with these columns for all states and territories.

states = [
'Alabama',
'Alaska',
'Arizona',
'Arkansas',
'California',
'Colorado',
'Connecticut',
'Delaware',
'District_of_Columbia',
'Florida',
'Georgia',
'Hawaii',
'Idaho',
'Illinois',
'Indiana',
'Iowa',
'Kansas',
'Kentucky',
'Louisiana',
'Maine',
'Maryland',
'Massachusetts',
'Michigan',
'Minnesota',
'Mississippi',
'Missouri',
'Montana',
'Nebraska',
'Nevada',
'New_Hampshire',
'New_Jersey',
'New_Mexico',
'New_York',
'North_Carolina',
'North_Dakota',
'Ohio',
'Oklahoma',
'Oregon',
'Pennsylvania',
'Rhode_Island',
'South_Carolina',
'South_Dakota',
'Tennessee',
'Texas',
'Utah',
'Vermont',
'Virginia',
'Washington',
'West_Virginia',
'Wisconsin',
'Wyoming'
]

subsets = []
for state in states:
    print(f"Processing {state}")
    df = read_state_2022_df(state)
    subsets.append(df[cols])

len(subsets)

# Export data

First the division-level data

df_2022 = pd.concat(subsets)

# Look at only Borough/County/Parish data
df_2022 = df_2022.dropna(subset='Division')

df_2022.head()

Let us make sure that there are differences

tdf1 = df_2020.set_index('FIPS')['Life Expectancy']
tdf2 = df_2022.set_index('FIPS')['Life Expectancy']
(tdf2 - tdf1).head()

df_2022.to_csv("../data/rwjf/division_2022.csv", index=False)

Then the state-level data

df_2022_states = pd.concat(subsets)
df_2022_states = df_2022_states[pd.isna(df_2022_states['Division'])]
df_2022_states.head()

tdf1 = df_2020_states.set_index('FIPS')['Life Expectancy']
tdf2 = df_2022_states.set_index('FIPS')['Life Expectancy']
(tdf2 - tdf1).head()

df_2022_states.to_csv("../data/rwjf/state_2022.csv", index=False)
